Human-readable config
Rules are now (recursively) parsed to rule objects as part of the ruleset loading.
This means that parsing is done only once per rule now instead of every match and validate.
This means a huge speedup during execution and gives us the opportunity to be a little more fancy and lenient with how the config should look.
Which in turn means config rules now read more natural:
- "Rule type" is now in the middle of the sentence: `amount greaterThan 0`
- You can use multiple words instead of camelCase: `amount greater than 0`
- Optional "is": `amount is greater than 0`, `description not empty`
- "not" goes after "is" when present: `description is not empty`

So we wave goodbye to RulesEngine, RulesMatcher, RulesValidator, and we say welcome to the entire Rules\Rules family.

Other changes
- Moved and renamed the config converter
- Extra exception handling during config loading
- Changed filename format for the new config files